Move show dialog when location is disabled code to another class
